requirements.txt
langgraph==1.0.5
langchain==1.2.3
langchain-openai==1.1.7
langchain-community==0.4.1
pydantic==2.12.5
pyyaml==6.0.3
openai==2.15.0
groq==1.0.0
sentence-transformers==5.2.0
vaderSentiment>=3.3.2
numpy>=2.4.1
pandas>=2.3.3
typing-extensions>=4.15.0
python-dotenv>=1.0.0
ipython>=8.0.0
